Synced

+++ alternate/create_nodes.py
@@ -9,12 +9,12 @@ file_lists=["node","node.o","nodeclass.o","util.o","nodeserver.o"]
 
 for i in range(arguments):
 	folder_name="Node"+str(i)
-	
-	try:
+
+	if(os.path.isdir(folder_name)):
 		for root, dirs, files in os.walk(folder_name):
 			for file in files:
 				os.remove(os.path.join(root, file))
-	except Exception:
+	else:
 	    os.mkdir(folder_name)
 
 	for file in file_lists:
